Automaton

+++ Automaton.java
@@ -95,8 +95,8 @@ public class Automaton {
     // using mapValToColor. This is how color is set generation to
     // generation.
     public static int mapAvgToVal(double avg, int kInt, String[] kCode){
-        double dif = kInt-avg;
-        double index = 3*(dif-1);
+        double dif = (double)kInt-avg;
+        double index = 3.0*(dif-1.0);
         int ind = (int)Math.round(index);
         return Integer.parseInt(kCode[ind]);
     }
